requirements.txt
beautifulsoup4==4.13.4
faiss-cpu==1.11.0
google-generativeai==0.8.5
langchain==0.3.26
langchain-community==0.3.27
langchain-huggingface==0.3.0
playwright==1.53.0
sentence-transformers==5.0.0
transformers==4.53.1
keybert==0.9.0
